+++ backend/src/main/java/com/ll/hotel/domain/member/member/controller/BusinessController.java
@@ -5,6 +5,7 @@ import com.ll.hotel.domain.member.member.dto.response.BusinessResponse;
 import com.ll.hotel.domain.member.member.entity.Business;
 import com.ll.hotel.domain.member.member.entity.Member;
 import com.ll.hotel.domain.member.member.service.BusinessService;
+import com.ll.hotel.domain.member.member.service.BusinessValidationService;
 import com.ll.hotel.global.rq.Rq;
 import com.ll.hotel.global.rsData.RsData;
 import jakarta.validation.Valid;
@@ -20,6 +21,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class BusinessController {
     private final BusinessService businessService;
+    private final BusinessValidationService businessValidationService;
     private final Rq rq;
 
     @PostMapping("/register")
@@ -27,7 +29,7 @@ public class BusinessController {
 
         Member member =rq.getActor();
 
-        String validationResult = businessService.validateBusiness(registrationInfo);
+        String validationResult = businessValidationService.validateBusiness(registrationInfo);
 
         Business business = businessService.register(registrationInfo, member, validationResult);
