+++ backend/src/main/java/com/ll/hotel/domain/member/member/service/BusinessValidationService.java
@@ -1,0 +1,54 @@
+package com.ll.hotel.domain.member.member.service;
+
+import com.ll.hotel.domain.member.member.dto.request.BusinessRequest;
+import com.ll.hotel.domain.member.member.dto.response.BusinessResponse;
+import com.ll.hotel.domain.member.member.type.BusinessApiProperties;
+import com.ll.hotel.global.exceptions.ServiceException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Service;
+import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.web.reactive.function.client.WebClientRequestException;
+import org.springframework.web.reactive.function.client.WebClientResponseException;
+
+import java.util.Map;
+
+@Service
+@RequiredArgsConstructor
+public class BusinessValidationService {
+    private final WebClient webClient;
+    private final BusinessApiProperties properties;
+
+    public String validateBusiness(BusinessRequest.RegistrationInfo registrationInfo) {
+        String apiUrl = properties.getValidationUrl() + "?serviceKey=" + properties.getServiceKey();
+        BusinessRequest.RegistrationApiForm registrationApiForm = BusinessRequest.RegistrationApiForm.from(registrationInfo);
+
+        try {
+            BusinessResponse.Verification response = webClient.post()
+                    .uri(apiUrl)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .bodyValue(registrationApiForm)
+                    .retrieve()
+                    .bodyToMono(BusinessResponse.Verification.class)
+                    .block();
+
+            if (response == null || response.data() == null) {
+                throw new ServiceException("500", "사업자 등록 검증 API 응답이 올바르지 않습니다.");
+            }
+
+            if (response.data().isEmpty()) {
+                throw new ServiceException("400", "사업자 정보가 조회되지 않았습니다.");
+            }
+
+            Map<String, Object> result = response.data().getFirst();
+            return (String) result.get("valid");
+
+        } catch (WebClientResponseException e) {
+            throw new ServiceException("402", "사업자 검증 API 호출 실패했습니다.");
+        } catch (WebClientRequestException e) {
+            throw new ServiceException("503", "사업자 검증 API 요청 중 네트워크 오류가 발생했습니다.");
+        } catch (Exception e) {
+            throw new ServiceException("500", "사업자 검증 중 예기치 못한 오류가 발생했습니다.");
+        }
+    }
+}
